requirement.txt
selenium 
beautifulsoup4 
undetected-chromedriver
requests
google-search-results
setuptools
python-dotenv